style : View 전용 Controller와 API controller
