+++ src/main/java/com/twentiethcenturygangsta/ourboard/controller/AdminAPIController.java
@@ -1,0 +1,48 @@
+package com.twentiethcenturygangsta.ourboard.controller;
+
+import com.twentiethcenturygangsta.ourboard.entity.OurBoardMember;
+import com.twentiethcenturygangsta.ourboard.form.LoginForm;
+import com.twentiethcenturygangsta.ourboard.manager.session.SessionConst;
+import com.twentiethcenturygangsta.ourboard.services.LoginService;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Controller;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+
+@Slf4j
+@Controller
+@RequestMapping("/our-board/api")
+@RequiredArgsConstructor
+public class AdminAPIController {
+
+    private final LoginService loginService;
+
+    @PostMapping("/login")
+    public String loginAPI(@ModelAttribute("loginForm") LoginForm loginForm, BindingResult bindingResult, HttpServletRequest request) {
+        log.info("bindingResult = {}", bindingResult);
+
+        // loginForm binding Error
+        if (bindingResult.hasErrors()) {
+            return "redirect:/our-board/admin/login";
+        }
+        OurBoardMember ourBoardMember = loginService.login(loginForm.getMemberId(), loginForm.getPassword());
+        log.info("ourBoardMember = {}", ourBoardMember);
+
+        // userName, password에 해당하지 않는 경우
+        if (ourBoardMember == null) {
+            bindingResult.reject("loginFail", "아이디 또는 비밀번호가 맞지 않습니다.");
+            return "redirect:/our-board/admin/login";
+        }
+
+        HttpSession session = request.getSession();
+        session.setAttribute(SessionConst.LOGIN_MEMBER, ourBoardMember);
+        return "redirect:/our-board/admin";
+    }
+
+}
